Moved the Editor into shared
