+++ src/MUNityClient/Shared/Los/LoSReaderComponent.razor
@@ -7,29 +7,7 @@
         <div class="col-12">
             <div class="card">
                 <div class="card-header">
-                    @if (ViewModel.SourceList.CurrentSpeaker != null)
-                    {
-                        <h5>Es spricht:</h5>
-                        <h5>@ViewModel.SourceList.CurrentSpeaker.Name</h5>
-
-                        @if (ViewModel.OutOfSpeakerTime)
-                        {
-                            <h6 class="text-danger">Bitte zum Ende kommen</h6>
-                        }
-                        else if (ViewModel.LowOnSpeakerTime)
-                        {
-                            <h6 class="text-danger">Verbleibende Zeit: @ViewModel.SourceList.RemainingSpeakerTime.ToString(@"mm\:ss")</h6>
-                        }
-                        else
-                        {
-                            <h6>Verbleibende Zeit: @ViewModel.SourceList.RemainingSpeakerTime.ToString(@"mm\:ss")</h6>
-                        }
-                    }
-                    else
-                    {
-                        <h5>Aktuell kein(e) RednerIn</h5>
-                        <h5>Redezeit: @ViewModel.SourceList.SpeakerTime.ToString(@"mm\:ss")</h5>
-                    }
+                    <LoSReaderHeader Mode="LoSReaderHeader.ReaderHeaderModes.Speaker" ViewModel="@ViewModel" />
                 </div>
                 <div class="card-body">
                     <h4>Redeliste</h4>
@@ -67,29 +45,7 @@
         <div class="col-12">
             <div class="card">
                 <div class="card-header">
-                    @if (ViewModel.SourceList.CurrentQuestion != null)
-                    {
-                        <h5>Es spricht:</h5>
-                        <h5>@ViewModel.SourceList.CurrentQuestion.Name</h5>
-
-                        @if (ViewModel.SourceList.RemainingQuestionTime.TotalSeconds <= 0)
-                        {
-                            <h6 class="text-danger">Bitte zum Ende kommen</h6>
-                        }
-                        else if (ViewModel.SourceList.RemainingQuestionTime.TotalSeconds <= 10)
-                        {
-                            <h6 class="text-danger">Verbleibende Zeit: @ViewModel.SourceList.RemainingQuestionTime.ToString(@"mm\:ss")</h6>
-                        }
-                        else
-                        {
-                            <h6>Verbleibende Zeit: @ViewModel.SourceList.RemainingQuestionTime.ToString(@"mm\:ss")</h6>
-                        }
-                    }
-                    else
-                    {
-                        <h5>Aktuell keine Frage oder Kurzbemerkung</h5>
-                        <h5>Redezeit: @ViewModel.SourceList.QuestionTime.ToString(@"mm\:ss")</h5>
-                    }
+                    <LoSReaderHeader Mode="LoSReaderHeader.ReaderHeaderModes.Question" ViewModel="@ViewModel" />
                 </div>
                 <div class="card-body">
                     <h4>Warteliste Fragen und Kurzbemerkungen</h4>
