+++ src/MUNityClient/Shared/Los/LosEditorSpeaker.razor
@@ -1,0 +1,126 @@
+﻿<div class="col-12 col-md-6 m-0 p-3">
+    <div class="row m-0 p-0 d-flex justify-content-between list-head border-bottom">
+        <div class="col p-0 m-0 mr-2">
+            <h5 class="col-12 p-0">
+                <i class="fas fa-user-alt mr-2"></i>
+                @if (ViewModel.SourceList.CurrentSpeaker != null)
+                {
+                    @ViewModel.SourceList.CurrentSpeaker.Name
+                }
+                else
+                {
+                    <span>-</span>
+                }
+                @if (ViewModel.SourceList.CurrentSpeaker != null)
+                {
+                    <span type="button" class="text-danger hover-enabled" title="Kopf leeren"
+                          @onclick="() => this.ViewModel.Handler.ClearCurrentSpeaker()">
+                        <i class="fas fa-trash ml-3"></i>
+                    </span>
+                }
+
+            </h5>
+        </div>
+        <div class="col-7 row m-0 p-0 d-flex justify-content-end">
+            <span type="button" class="py-1 px-3 text-danger hover-enabled"
+                  @onclick="() => this.ViewModel.Handler.AddSpeakerSeconds(-10)">-10</span>
+            @if (this.ViewModel.SourceList.CurrentSpeaker != null)
+            {
+                <h3 class="m-0 text-center @(ViewModel.LowOnSpeakerTime ? "text-danger" : "")">
+                    @(ViewModel.OutOfSpeakerTime ? "-" : "")
+                    @ViewModel.SourceList.RemainingSpeakerTime.ToString(@"mm\:ss")
+                </h3>
+            }
+            else
+            {
+                <h3>@ViewModel.SourceList.SpeakerTime.ToString(@"mm\:ss")</h3>
+            }
+
+            <span type="button" class="py-1 px-3 text-success hover-enabled"
+                  @onclick="() => this.ViewModel.Handler.AddSpeakerSeconds(10)">+10</span>
+        </div>
+    </div>
+    @* Controls *@
+    <div class="col-12 p-0 my-3">
+        @if (ViewModel.SourceList.Status == MUNity.Models.ListOfSpeakers.ListOfSpeakers.EStatus.Speaking 
+            || ViewModel.SourceList.Status == MUNity.Models.ListOfSpeakers.ListOfSpeakers.EStatus.Answer)
+        {
+            <i title="Rede pausieren" type="button" class="fas mx-2 fa-lg fa-pause"
+               @onclick="() => ViewModel.Handler.Pause()"></i>
+        }
+        else
+        {
+            <i title="Rede fortsetzen" type="button" class="fas mx-2 fa-lg fa-play"
+               @onclick="() => ViewModel.Handler.ResumeSpeaker()"></i>
+        }
+        <i title="Nächste Rede" type="button" class="fas mx-2 fa-lg fa-step-forward"
+           @onclick="() => ViewModel.Handler.NextSpeaker()"></i>
+        <i title="Redezeit zurücksetzen" type="button" class="fas fa-lg fa-undo ml-3"
+           @onclick="() => ViewModel.Handler.ResetSpeakerTime()"></i>
+        <i title="Antwort starten" type="button" class="fas mx-2 fa-lg fa-redo"
+           @onclick="() => ViewModel.Handler.StartAnswer()"></i>
+    </div>
+
+    <div>
+        <h5>Warteliste</h5>
+        @foreach (var speaker in ViewModel.SourceList.Speakers)
+        {
+            <div class="card border p-2">@speaker.Name <i class="fas fa-times mr-2 text-danger" @onclick="() => RemoveSpeaker(speaker)"></i></div>
+        }
+
+
+        @if (ViewModel.SourceList.ListClosed)
+        {
+            <button @onclick="() => ViewModel.Handler.OpenSpeakers()" class="btn btn-outline-success">
+                Redeliste
+                öffnen
+            </button>
+        }
+        else
+        {
+            <div type="button" @onclick="() => ViewModel.Handler.CloseSpeakers()" class="border mx-0 mt-2 p-2">
+                <i class="fas fa-ban mr-2 text-danger"></i>Redeliste
+                schließen
+            </div>
+        }
+    </div>
+    <div class="col-12 row mx-0 px-0 my-3">
+        <EditForm Model="NewSpeaker" OnValidSubmit="AddToSpeakers()">
+            <DataAnnotationsValidator />
+            <ValidationSummary />
+            <div class="col-11 py-0 pl-0">
+
+                <div class="form-group">
+                    <InputText @bind-Value="NewSpeaker.Name" class="form-control" placeholder="Land hinzufügen..." disabled="@ViewModel.SourceList.ListClosed" />
+                </div>
+
+            </div>
+            <div class="col-1 p-0">
+                <button type="submit" class="btn btn-success m-0 py-2 px-0 w-100" disabled="@ViewModel.SourceList.ListClosed"><i class="fas fa-plus fa-lg"></i></button>
+            </div>
+        </EditForm>
+    </div>
+</div>
+
+@code {
+    [Parameter]
+    public ViewModels.ListOfSpeakerViewModel ViewModel { get; set; }
+
+    private Models.ListOfSpeaker.SpeakerToAdd NewSpeaker = new Models.ListOfSpeaker.SpeakerToAdd();
+
+    private void RemoveSpeaker(MUNity.Models.ListOfSpeakers.Speaker speaker)
+    {
+        this.ViewModel.Handler.Remove(speaker);
+    }
+
+    private EventCallback AddToSpeakers()
+    {
+        if (NewSpeaker.Name != null && NewSpeaker.Name.Length > 2)
+        {
+            this.ViewModel.Handler.AddSpeaker(NewSpeaker);
+            NewSpeaker.Name = "";
+        }
+
+        return EventCallback.Empty;
+    }
+}
